pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docsiphon"
version = "0.1.1"
description = "Turn documentation sites into AI-ready local Markdown corpora with preserved paths and audit artifacts."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Terry" }]
keywords = [
  "ai",
  "rag",
  "documentation",
  "markdown",
  "crawler",
  "sitemap",
  "llms.txt",
  "knowledge-base",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "requests>=2.32.5",
  "chardet<6",
  "beautifulsoup4>=4.12.2",
  "markdownify>=0.12.1",
  "readability-lxml>=0.8.1",
  "trafilatura>=1.9.0",
  "tomli>=2.0.1",
]

[project.scripts]
docsiphon = "docsiphon.cli:main"

[dependency-groups]
dev = [
  "pre-commit>=4.3.0",
  "pytest>=8.4.2",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
